Fix bugs with locks imposed upon vectors in VecScatterBegin/End
Athough otherwise one can potentially overlap computation/communication in MatMultTranspose_MPIAIJ,
MatMultTransposeAdd_MPIAIJ, that is based on the assumption that the scattered data is not written
to the destination vector until VecScatterEnd(). This assumption is too strict for implementation
of VecScatter and could lead to wrong results. For example, the origin MatMultTransposeAdd_MPIAIJ is
wrong with -vecscatter_merged. In addition, the assumption is not stated in petsc user's manual.