Don't log to the configure file language pushes and pops
The logging produces a lot of gibberish (especially at the top of the log file; but even in places
in the middle of the log file) where it repeatedly logs pushing and poping but shows no other
intervening activity.

Since the name of the compiler is shown for each compile test it is possible to track which
language is being tested without the push and pop logging. In all the years of debugging configure
with configure.log I don't think I've used the push and pop messages to help understand the problem

As a side benefit, on a basic ./configure on my Mac this cut the size of the log file by 9.5 percent

Commit-type: style-fix, feature