Introduce logging of memory usage to the PETSc event logging infrastructure.
The logging can be turned on and viewed with the options -log_view -log_view_memory

The logging requires the use of the PETSc malloc so may slow the runs slightly but since it is not
doing error checking of the memory it should not slow down the computations significently.

Four pieces of information about memory are logged and display in four additional columns in the table.
They are documented in the output above the table. Each column provides a particular "view" of the
memory usage in the event and they complement each other.

This should make it much easier for developers focusing on memory usage in their implementations
to understand the usage in the code, where the memory is being utilized, how it scales with problem size and number of processes etc.

Commit-type: feature