DMDA VTK Viewing: for multi-dof DMDAs with named fields, output multiple scalar arrays
When viewing vectors associated with a DMDA, restore the previous
behavior, with respect to ea2d7708fa6bebc714ed9a4ec0380dd837201800,
of outputting a separate scalar field for each component IFF any
component is named. Otherwise, retain the current behavior of outputting
a single vector-valued field.

The rationale is that if the components are unnamed, the user likely
considers them to be vector data. On the other hand, if any are named,
then the user likely would like to deal with them as separate fields.

Add additional logic to grvtk.c to detect if any DMDA fields are named,
using a static helper function DMDAGetFieldsNamed (but still exacerbating the
existing code duplication).

Field names are "VectorName[.FieldName]", and default names are e.g. "Unnamed.2".

Change the size of the "array2" values to be large enough. This is
potentially wasteful if not using named fields.

Add additional DM test ex48 tests to the test suite.

Add a note to the DMDAVTKWriteAll man page.

Reported-by: Jed Brown <jed@jedbrown.org>
Thanks-to: Dave A. May <dave.mayhem23@gmail.com>