DMPlex: Speedup DMPlexStratify()
The depth label encodes a (non overlaping) partition of the points in the
Plex chart. The set of points in each level can be represented with a
contiguous range of integer. From these assumptions, the depth label can
be constructed very quickly and using negligible memory by storing the
label strata as stride index sets.