Fixed inconsistent naming with CG methods for TR
Some of the KSPXXX macros for xxx had an inconsisten KSPCGXXX that was not in the string xxx name
The KSPCGXXX was used for KSPCreate_CGXXX() but the other method functions were missing the CG, for example,  KSPSolve_XXX()
   (these meant using tags to find functions was cumbersome due to the special cases)
Missing Fortran macro

Removed the extraneous CG from the macros and the function usage; now all macros and functions have consistent PETSc style usage
   that match the string names
Added deprecation messages for out-dated use
Added new configure check for PETSC_DEPRECATED_MACRO

I decided to remove the CG from the KSPCGXXX instead of adding a cg to the string name because 1) it was not possible to deprecate
the string names at compile time, we would have to have runtime warnings. 2) the string names are shorter, 3) the stcg already
had a cg (at the end and not the beginning) and that would have to be removed or we'd have the absurd cgstcg. If starting from
sratch today not worrying about legacy I would have all the string names begin with cg consistently (and naturally the KSPXXX would
have them as well), but for legacy reasons we have the current naming which is not ideal because it does not clearly
indicate which methods are CG methods. (We have this problem with other subclasses in PETSc also).

Commit-type: style-fix