DMPlex: Update Gmsh reader to set hybrid bounds by default
If the reader is asked to return an interpolated mesh and the mesh
has tri+quad or tetra+wedge cells, you most likely want the output
DMPlex to have the hybrid bounds set, so make it default behaviour.

If for whatever reason the users do not want the hybrid bounds set,
they can still use `-dm_plex_gmsh_hybrid 0`.