Add DMStagDuplicateWithoutSetup()
Creates a copy of the DM, prior to setup.
This should allow this function to be used within DMClone_Stag(), DMCoarsen_Stag(),
DMRefine_Stag(), and DMStagCreateCompatibleDMStag().

Note that DMClone() serves a similar but different purpose. It is
intended to create a copy of a DM with the same topology, which
is set up if the original DM is set up.