Man pages: remove .keywords: fields
This field is recognized by doctext from Sowing (http://wgropp.cs.illinois.edu/projects/software/sowing/doctext/node20.htm#Node20),
but it doesn't seem to be uniformly used or maintained.

Thus, remove all .keywords: fields, and a following blank line, if present.

This is accomplished with GNU sed (gsed on OS X), with the following commands.
*Warning* that this type of command can corrupt a .git directory,
so be cautious in reusing or modifying these commands. They first look
for and delete matching lines with a following line consisting of only whitespace,
and then delete any remaining matching lines.

    find src/ -type f -exec gsed -i '/keywords:/ {N; /\n\s*$/d}' {} +
    find src/ -type f -exec gsed -i '/keywords:/d' {} +
    find include/ -type f -exec gsed -i '/keywords:/ {N; /\n\s*$/d}' {} +
    find include/ -type f -exec gsed -i '/keywords:/d' {} +

Hints on the sed command obtained from: https://unix.stackexchange.com/questions/100754/how-to-delete-a-specific-line-and-the-following-blank-line-using-gnu-sed