Man pages: remove Concepts: fields
These fields were previously stripped from the man pages by logic removed in 
21a59cba2737d49dc2f0bd12c08db0d2a3f3f209

Remove these fields from all man pages (but not from examples).

This is accomplished with GNU sed (gsed on OS X), with the following commands.
*Warning* that this type of command can corrupt a .git directory,
so be cautious in reusing or modifying these commands. They first look
for and delete matching lines with a following line consisting of only whitespace,
and then delete any remaining matching lines.

    find src -type f -not -path "*/examples/*" -not -name "*.html"  -not -name "*.bib" -exec gsed -i '/Concepts:/ {N; /\n\s*$/d}' {} +
    find src -type f -not -path "*/examples/*" -not -name "*.html"  -not -name "*.bib" -exec gsed -i '/Concepts:/d' {} +
    find include -type f -not -path "*/examples/*" -not -name "*.html"  -not -name "*.bib" -exec gsed -i '/Concepts:/ {N; /\n\s*$/d}' {} +
    find include -type f -not -path "*/examples/*" -not -name "*.html"  -not -name "*.bib" -exec gsed -i '/Concepts:/d' {} +

Hints on the sed command obtained from: https://unix.stackexchange.com/questions/100754/how-to-delete-a-specific-line-and-the-following-blank-line-using-gnu-sed