Have BlasLapack search newer MKL install locations before old
Add search for GNU thread based MKL libraries (for use with Gnu compilers).
On compute002 using the threaded ddot() scales very nicely with the number of
cores unlike the results on Barry's Mac laptop reported in a previous commit

Note that will will ignore MKL libraries that use TBB threading and only use
ones that use OpenMP threading

Note: the current directory search for MKL libraries is ad hoc and disorganized because it contains
a legacy of older MKL directories. This change tries to have it look in more recent locations for MKL
but it may still at times detect older styles of MKL and accept them before checking newer versions. This
may be problematic for finding MKL libraries that use threading