DMStag: correct number of ghost elements for ghosted boundaries with zero stencil width
In the (not practically very useful) case of ghosted boundaries and zero stencil width, fix a bug in calculating the number of elements in the local representation. In this case (only) an extra partial dummy element needs to be added, relative to the periodic case.

Add tests with DMStag test ex1.