DMStag: Add support for explicit injective local-to-global map
Previously, one could not use periodic boundary conditions in a direction with only one rank. This was because the usual method to create a 1-1 (injective) local-to-global map is to reverse the (1-to-many) global-to-local map, disregarding any off-rank communication. This fails in the 1-rank-periodic case since several rank-local local dof map to the same rank-local global dof.

Here, a new developer method is added to construct an
explicit local-to-global map. This is called during
DMSetUp_Stag() in the case mentioned above.

DMLocalToGlobalBegin_Stag() and DMLocalToGlobalEnd_Stag() use this explicit local-to-global mapping if it exists,
and otherwise attempt the previous behavior.