DMPlex MED: fix types to use med_idt and add detail to error messages
MED-4 tags the upper bits of the med_idt (a 64-bit type), thus does
not work with PetscInt, which could be 32-bit. The failure symptoms
are obscure, just returning that there are no meshes instead of
indicating that the file ID is invalid.  This commit uses accurate
types.

Commit-type: bug-fix