SF: Partially fix a bug when overlapped SF communications have same rootdata or leafdata on some ranks
Now we use two keys (rootdata, leafdata) to identify a pending communication. But that is still not
enough for cases where communications have same rootdata and leafdata on some ranks. Currently We
error out on these cases. See ex2.c for various cases we can handle and we can not handle.