DMStag: add DMStagCreateISFromStencils(), and test ex17
This allows one to create an IS giving indices
into a global vector associated with a DMStag,
in terms of a set of DMStagStencil objects. The
element indices (i,j,k) are ignored, but the location (loc)
and component (c) entries are used to define a subset of the
dof.

This is useful e.g. for extracting a submatrix for testing purposes,
and may also be useful for use with PCFieldSplit.

Also adds tests.