Support viewing vecs with multifields in VTK from DMPlex
Change precedence when finding sections in DMPletscVTKWriteALL

Searching for "section" before resorting to DM makes it easier to view just a single
field of a vector: rather than computing a subdm, we just get the field section and
compose it with a thin duplicate of the vector

Use composed "section" in DMPlexVTKWriteAll_VTU

Handle multifield vecs in VecView_Plex_Local_VTK

Test VTK multifield output in ascii and vtu in snes tutorials ex24

Make plexvtu use the sections composed with the vectors, provide vector output, and have more sensible names

Tranfer vector fields expanded to 3 components in one transfer

Only put a number on a field if there is more than one field

Add fieldnum argument to PetscViewerVTKAddField()

Now a vector can be added only for a single field to be printed.  We can add the same field
multiple times.  This avoids copying fields of a vector into separate vectors for viewing