Increase err verbosity
In commit 984117c42c7 a fix was put in to handle cases when stderr
goes to stdout; however, it seems that sometimes we have cases where
mpi_abort messages go to stdout but the petsc trace goes to stdout.
Here, we cat both stdout and stderr with stdout first.  This might be
too verbose, but worth gaining experience.