Make it easier to test the last failed tests
Re-running the last failed tests were a pain because of how loops and
subtests were labelled with a suffix that started with "_" as a
delimiter.  Because the underscore is commonly used by developers, this
created a non-unique mapping from test name to target name.  The first
fix is to add in a delimiter for these cases ('+') so that unique
mapping from testname to target name can be created.

This can result in long string of targets to run, so report_tests.py
creates a silly little script to do nothing but echo the tests that
failed.  A new search-like argument 'test-fail=1' can be invoked to
re-run all of the failed tests.  The new syntax is:
     make -f gmakefile test test-fail=1
To see which tests would be run:
     make -f gmakefile print-test test-fail=1