Enable rich querying of tests
Tool for querying the tests.

Which tests to query?  Two options:
    1. Query only the tests that are run for a given configuration.
    2. Query all of the test files in the source directory
For #1:
   Use dataDict as written out by gmakegentest.py in $PETSC_ARCH/$TESTBASE
For #2:
   Walk the entire tree parsing the files as we go along using testparse.
   The tree walker is simpler than what is in gmakegentest.py because it
   can avoid all of the logic regarding configuration.
See documentation at the top of query_tests.py for further details.
The dataDict is further transformed to allow fast searching.

The query_test.py script by default outputs the info into a form that is
able to be used by gmakefile.test to enable this functionality:
  make -f ${makefile} test query='requires' queryval='*MPI_PROCESS_SHARED_MEMORY*'

More rich querying can be done from within ipython.  This is documented
in the developer's documentation.