Move to new filter/filter_output method
Change to using filters as arguments to petscdiff.  By moving the
manipulation of the input and output into petscdiff, this fixes problems
in using piped commands (based on original, ancient test system), which
is problematic in capturing the right error codes.  This also required
fixes for some tests: the "filter: Error: ..." is used to test stderr of
tests.

This fixes problems with REPLACE=1 as reported by @knepley and debugged by @sbalay.