Improvements to artifacts
Following the gitlab ref page on artifacts, this is an attempt to
upload the test log files and the gcov tarbal from `make gcov`
Includes update of gcov.py to use optparse and eliminate os.environ for
getting petsc_dir and petcs_arch.
Also include {configure,make}.log to artifacts.