DMNetwork: Setting/accessing variables for components.
DMNetwork till now only allowed setting/accessing the variables at a vertex or edge. For vertices/edges that have multiple components (which presumably could represent different physics) the user would need to some book-keeping to keep track of variable offsets associated with a component. This feature avoids the need of this book-keeping by allowing setting variables for components and managing the variable offsets for them.