DMNetwork: Added example for testing component variables feature.
This example is a modification of ex9busdmnetwork.c that has the following changes.
i) A separate struct 'Exc' is used for managing the exciter component. This data was a part of 'Gen' struct.
ii) Instead of setting all the variables at the vertex via DMNetworkAddNumVariables(), the
    bus, gen, and exc components set the number of variables associated with them via
    DMNetworkSetComponentNumVariables().
iii) Instead of accessing the variable offset for a vertex via DMNetworkGetVariableOffset(), the
     bus, gen, and exc components retrieve their variable offsets via DMNetworkGetComponentVariableOffset().
     This allows more granularity and thereby avoids keeping track of the offsets as done in ex9busdmnetwork.c