TS: add support for transient variables in BDF
This feature enables conservative time integration even when using
nonconservative state variables.  The formulation, which requires only
the definition of transient (conservative) variables C(U) as a
function of the state variable U, is appropriate for linear multistep
methods for ODE and DAE.

This technique could be extended to Runge-Kutta methods by creating an
index-1 DAE by augmentation with the equation C = C(U).  The stage
equations in diagonally implicit methods generally ask to solve for Y:

  F(Y, alpha Y + Z) = 0

where Y = [C, U] and Z is known from prior stages.  With the augmented
equation, this decouples to

  C - C(U) = 0
  f(U, alpha C + C_z) = 0

This commit includes an example demonstrating loss of conservation in
chain rule primitive-variable IFunction, while preserving it for
pure-conservative variables and the transient variable formulation.

Resolves #547

Requested-by: Gautam Bisht <gautam.bisht@pnnl.gov>
Thanks-to: Hong Zhang <hongzhang@anl.gov>