DMPlexGetIndicesPointFields_Internal: add islocal argument and docs
This fixes a serious bug in DMPlexGetClosureIndices (and likely many
other places) that returns corrupt local indices in case a constrained
component is not last.  See example in new comment for
DMPlexGetIndicesPointFields_Internal.  The prior behavior was correct
for global sections, but many/all uses were with local sections.

The same bug applies to DMPlexGetIndicesPoint_Internal, which is also
updated in this commit.

Thanks-to: Matthew Knepley <knepley@gmail.com>