PC: ASM, Block Jacobi: when viewing, note local solver on rank 0 is shown
The KSPView() call shows the (sub-)KSP on rank 0, which should have the same
type, PC, parameters, etc., but will not be the identical solver,
as the size of the local subdomains and the entries of the local
submatrices will be different. For instance, the output only shows
the local Mat for rank 0, which might confuse a user into thinking that
the subdomains are all of the same size when they are not. Similarly,
a PC like LU would show different factorization details on different
ranks, even though the solver settings are the same.

Make the viewer output more precise by noting explicitly that is
displaying information for the rank 0 subsolver.

Update test output with the new string.