Several improvements to pinned memory support for VECCUDA.
This commit does the following:

* Moves minimum_bytes_pinned_memory out of the Vec_CUDA structure and makes it a
member of struct _p_Vec; initializes it to 0 at vector creation time.

* Adds a flag inside _p_Vec indicating whether there is a current host (CPU)
allocation that is using pinned memory. This flag is used to determine the
appropriate free() implementation to use for a host allocation.

The above changes make it possible and safe to use at any time the API for
setting the minimum size for the use of pinned memory for host allocations.

VecCUDASetPinnedMemoryMin()/VecCUDAGetPinnedMemoryMin() becomes
VecSetPinnedMemoryMin()/VecGetPinnedMemoryMin() with these changes.
These interfaces now exist for VECVIENNACL as well, but have no effect, as
support for pinned host allocations needs to be added for this case.