Docs: add Sphinx docs
Platform on which to add Sphinx documentation.

Since this exists in parallel with the current documentation system,
confine additions to src/docs/sphinx_docs, with the exception
of a .readthedocs.yml file in the root.

One should be able to generate HTML documentation locally and
on ReadTheDocs.

Includes an extension (not used in this commit) to add inline links
to HTML documentation.