Fixing the too many args problem
PETSc now has so many tests that the globsearch which used a python
one-liner with the testlist coming in from stdin was failing.  This
upgrades the query_tests.py script to enable doing the glob-style
searching on just the names.

make: use $(info) instead of echo to work with large line lengths