gcc/gfortran-10: fix errors/warnings
The primary issue now is - gfortran defaults to flagging argument mismatch as error.
We can work-around this with '-fallow-argument-mismatch' option - however this switches from errors to warnings.

And the only way to disable this warning is '-w' i.e disable all warnings.

Fixes involve adding interface definitions for all prototypes that get used. [and work-around mpich by using a(1) for scalars]