Enable HIP and SYCL configuration
This adds support for HIP and SYCL configuration.

HIP is both a compiler (hipcc) and a package to be searched for
(hipblas, hipsparse) similar to CUDA.
Language:  HIP; Compiler: HIPCC (even though it is C++)

SYCL is currently implemented for searching only the dpcpp compiler.
Because Intel's oneAPI is very similar to MKL, Blas/Lapack is handled by
those packages.
Language:  SYCL; Compiler: SYCLCXX

Although it is a C++ compiler, both hipcc and dpcpp are very slow.
Currently using the .hip.cpp and .sycl.cxx extensions although more work
will be needed for testings.  See the kokkos work.

This includes preliminary changes for pipeline testing
The HIP_PLATFORM is an environment variable that chooses the backend.
So one configuration file, but the platform is specified in
.gitlab-ci.yml.

This includes beta versions of configuration files for what could be used on
Frontier.  This is both the HIP version and CUDA versions for enabling
comparisons.