Add target to top-level Makefile to build Sphinx docs in a virtual env.
As a convenience, add a target to the top-level makefile which creates
a virtual environment using Python 3.3 or later (so that the venv module
exists by default).

Also add a target to clean these docs, cleaning up using the Sphinx
makefile, and also removing the virtual environment created.

For many developers, it would be more natural to make sure that the
requirements in src/sphinx_docs/requirements.txt are satisfied with
their local Python, and use work in src/sphinx_docs, using the Makefile
there.

Note: the Sphinx build itself will still configure its own minimal build of
the "classic" PETSc docs (to have an index of man pages available),
using a dedicated PETSC_ARCH (currently arch-sphinx-minimal).
This is redundant and confusing in some ways, but has the
advantage of having the Sphinx build be fairly self-contained, which is
very handy for building it e.g. on ReadTheDocs, or indeed in an
unconfigured local build of PETSc.