Mark MKL matrix handle as not optimized after MatProduct operations.
I had previously incorrectly assumed that the matrix optimization step is
carried out by MKL at the end of a matrix-matrix product operation, but
a conversation with the MKL team confirms that this is not the case.
This did not cause a correctness issue, but might have led to reduced
performance if the matrix was then used for repeated operations that could
benefit from the optimization step.

Thanks-to: Maria Zhukova <maria.zhukova@intel.com>