Add a filter for stderr output caused by misbehaving Intel MPI wrappers.
The root of the issue is that Intel's MPI team keeps assuming that
/bin/sh invokes Bash, but on some systems this invokes a shell that is
only guaranteed to have POSIX features. Ubuntu is one such system. This
leads to output to stderr from the Intel MPI compiler wrappers due to
use of Bash features in the scripts, though the wrapped compilers are
actually producing correct output.