Docs: override HTML translator dirhtml and readthedocsdirhtml builders
This allows our extension to insert man page links with the dirhtml
builder, as it does with the html builder. The correspnding change
is also made for ReadTheDocs' custom version of the dirhtml builder.
The same danger is present in doing this (they are free to change
the name of that builder and break our link generation), but note
that a great deal of information is available on the build on
ReadTheDocs (which is how the name of the readthedocsdirhtml builder
was determined).