Manual: add custom first and last page to LaTeX version
This uses the same LaTeX functionality as the former LaTeX manual,
inserting the custom LaTeX source with the "latex_elements" variable
in conf.py for the Sphinx docs.

"make manual_pdf" for the Sphinx docs should now generate a PDF which,
like the old manual, has the correct pages at the beginning and end to
qualify as an ANL Tech Report.