Docs: Sphinx docs manual build: simplify logic
Use the "latex_documents" option to directly specify a subset of the
documents to use for the latex builder. This allows the removal of
custom logic to change the root directory for the manual, remove al of the separate
output directory for the manual, and allows use of the "latexpdf" make mode.