Docs: remove old LaTeX-only manual
The manual is now exclusively built with Sphinx, both as part of the
Sphinx HTML docs, and as a standalone PDF built from a subset of the
Sphinx source, using the Sphinx latex builder.

Remove several images and figures which apparently aren't used
anywhere, except for those which are source for images still used
in the manual, which are moved into the Sphinx docs source subtree.

BibTeX files are left in place, as they are used by the Sphinx
docs as well as other scripts (and perhaps even people writing papers).

The TAO LaTeX-only manual is unaffected by these changes.