Fix pcmgsetlevels() for Fortran to properly convert the C MPI communictors to Fortran
Fix PCMGSetLevels_MG() to support the use of MPI_COMM_NULL for communicators that that live
on a subset of the orginal communicator.

Add additional documentation material to PCMGSetLevels() explaining how one must handle the restriction and interpolation
operations that move between different levels which have a different number of MPI ranks.

Commit-type: bug-fix, docs
/spend 2h
Reported-by: Shatrov Oleg <shatrov.oleg.a@gmail.com>