The previous code passed options for the PETSc C, C preprocessor and CXX compiler to the ccbin compiler used by nvcc even if it was a completely different compiler and the arguments caused the compiler to error with unknown arguments.
This has been an outstanding problem for years, this is the first crude attempt to fix it by
not passing the arguments unless the compiler is identical to the PETSc compiler

TODO: have a better check that the compilers are the same
TODO: provide a systematic way for the user to provide the compiler used by nvcc, perhaps default to the compiler used by PETSc?

/spend 3.5h
Thanks-to: Jose E. Roman <jroman@dsic.upv.es>

Fix typo in list of c99 flags

Add error message regarding Intel compilers breaking Gnu compilers when c99 test fails with gcc.

Eventually we may want more systematic checks of Intel compilers breaking Gnu compilers

/spend 15m