ComponentName: modification of DMPlexCreateGlobalToNaturalSF
The modification of DMPlexCreateGlobaltoNaturalSF allows to pass NULL as an argument
for the sequential section when calling the routine. The interface hasn't change. It takes
into account that a DM can be distributed with DMPlexDistribute just after its creation.
In that case, sfNatural cannot be computed in DMPlexDistribute since no section is defined on the DM
at that moment. The user should then call DMPlexCreateGlobaltoNaturalSF after a section is set to the
parallel DM in order to compute the sfNatural. Instead of computing the sequential section needed for
DMPlexCreateGlobaltoNaturalSF, he can now pass NULL as an argument for the section and it is reconstructed
from the parallel section in the routine.
Error messages are slightly modified in all routines of plexnatural.c when the naturalSF is not present and
useNatural is set to TRUE. It mentioned now that a section has to be defined for the naturalSF to be computed.
TODO: since DMPlexCreateGlobaltoNaturalSF is not typically called by the user, automatically call it as soon
as useNatural is TRUE, the DM has been distributed and a section is present.
No failed test when running all dm tests locally after the modification.

Reported-by: Blaise Bourdin <bourdin@lsu.edu>
Thanks-to: Blaise Bourdin <bourdin@lsu.edu>