Fix pullback matrix indexing in dspacelagrange.c
Thanks to: Yang Zongze

I think the incorrect indexing wasn't causing problems before because it doesn't matter if pNk == 1.
That's the case when the form degree we are pulling back is 0 or the boundary points' dimension.  So
If we're only testing d <= 3, this should only have cause problems pulling a 1-form back onto a 2D face.
I will have to check to see if we have a regression test for this case.