Docs: Build some "classic" docs with Sphinx docs
* Move the minimal configure of PETSc to its own script
* Add logic in the Sphinx conf.py to run this script to
  * configure PETSc if needbe and build some docs
  * move them into a location to be included with html_extra_path

This allows pages generated with the "classic" logic to be used
consistently with Sphinx builds.

This removes the ability to steal htmlmap from an existing
docs build, because the required html sources are in the src/ and
include/ tree and we hope to easily copy them.

To avoid having to wait around too much, the "classic" build won't auto-update with changes
(you are instructed with a message what to delete to do this).

Includes some custom logic to try to detect if being build on
GitLab CI, to set base URLs for man page links correctly.

Includes some placeholder logic for ReadTheDocs, to be removed.