CI/Docs: Update GitLab pipelines for docs review and deployment
Rename the current .test job to .test-basic, which defines the general conditions
under which testing/review pipeline jobs should run.

Re-add .test to inherit from this and add an additional condition which
skips it if a "docs-only" label is present on an MR.

Change check-ci-settings and pause-for-approval to depend on
.test-basic, so they are not skipped, even for a docs-only run.

Create a new hidden job .docs in stage-2 which prepares the environment for a
Sphinx docs build on our dedicated runner. Note the special tag
and particular invocation of sphinx-build.

Change .docs-review to depend on .docs as well as .test-basic.

Remove the .docs-only-review job.

Add a new "pages" job to deploy to GitLab pages. This is currently
the only job which does not depend on .test-basic, because it is
not for testing/review, but for deployment, when particular branches
are updated.

The deployment is intended to produce
- public/main from branch main
- public/release from branch release
- public/x.y from branch release-x.y
- testdocs* from branch testdocs*